requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlmodel==0.0.20
asyncpg==0.30.0
psycopg2-binary==2.9.11
alembic==1.14.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
Pillow==11.0.0
python-dotenv==1.0.1
websockets==13.1
pydantic-settings==2.13.1
email-validator==2.3.0
